Fix rounding issue
Fixes a bug present when the amount value has more than two decimal places e.g:

Product with price set to £35.95
Discount of 15% applied, setting the price to 30.5575
Magento rounds this up to £30.56, but this plugin multiplies 30.5575 by 100 to get the pence value without rounding first. This means the 'amount' value is set to 3055.75 and the JSON APi throws an error because the 'amount' isn't an integer.